changed file and folder name + fixed dlib initlization
