+++ DlibLipsPredictor/dlibSupportFunctions.py
@@ -10,8 +10,6 @@ import dlib
 def resizeImage(img, dim=(650, 650)):
     resized = cv2.resize(img, dim, interpolation=cv2.INTER_AREA)
     return resized
-
-
 # ----------------------------------------------------------------------------
 # function used to intilize dlib objects
 # input: none
@@ -20,7 +18,6 @@ def initializeDlib():
     detector = dlib.get_frontal_face_detector()
     predictor = dlib.shape_predictor("../dlib-predictor.dat")
     return detector, predictor
-
 # ----------------------------------------------------------------------------
 # function used to rotate image
 # input: image
@@ -28,3 +25,20 @@ def initializeDlib():
 def rotateImage(img):
     rotated = cv2.rotate(img, cv2.ROTATE_90_COUNTERCLOCKWISE)
     return rotated
+# ----------------------------------------------------------------------------
+# function used to exctract frames from video
+# input: video
+# output: frames
+def getVideoFrames(videoPath):
+    """Function to return a video's frames in a list
+    :type videoPath: String
+    """
+    vidcap = cv2.VideoCapture(videoPath)
+    if vidcap.isOpened() == False:
+        return [], False
+    success, image = vidcap.read()
+    allFrames = []
+    while success:
+        allFrames.append(image)
+        success, image = vidcap.read()
+    return allFrames, True
